refactor modals property to navigation.modals
