+++ src/client/src/containers/RightSidebar/index.tsx
@@ -17,7 +17,7 @@ import messages from "./strings";
 import { ReactComponent as Cancel } from "../../assets/cancel.svg";
 
 import { AppState } from "../../store/combineReducers";
-import * as ModalActions from "../../store/modals/action";
+import * as ModalActions from "../../store/navigation/modals/action";
 import { hasServicesSelector } from "../../store/azureProfileData/servicesSelector";
 import { getIsVisitedRoutesSelector, IVisitedPages } from "../../store/templates/preview/wizardNavigationSelector";
 import ProjectDetails from "./ProjectDetails";
