+++ src/client/src/store/navigation/combineReducers.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux";
+import modals from "./modals/combineReducers";
+
+const navigationReducer = combineReducers({
+  modals
+});
+
+export default navigationReducer;
+export type ConfigState = ReturnType<typeof navigationReducer>;
